backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pillow==10.0.0
requests==2.31.0
eth-account==0.8.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==1.10.9
